Refactored code from eip-client to eip-commons
