+++ commons/src/main/java/com/ozonehis/eip/utils/ShutdownHandler.java
@@ -1,4 +1,11 @@
-package com.ozonehis.eip;
+/*
+ * Copyright © 2021, Ozone HIS <info@ozone-his.com>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.ozonehis.eip.utils;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
